Add tests for datapackage check and init + small refactoring
